refactor: replace platform IAM functions with internal handlers for user and role management
